Added I2C timeout. Moved some code into the common crate.
