+++ tm4c-hal/src/lib.rs
@@ -11,6 +11,7 @@ extern crate nb;
 pub mod bb;
 pub mod delay;
 pub mod gpio;
+pub mod i2c;
 pub mod serial;
 pub mod sysctl;
 pub mod time;
